Refactor Checkbox and Radio
