--- src/checkbox/index.js
@@ -1,3 +1,0 @@
-import Checkbox from './src/components/Checkbox';
-
-export default Checkbox;
