+++ src/choices/index.js
@@ -1,0 +1,7 @@
+import Checkbox from './src/components/Checkbox';
+import Radio from './src/components/Radio';
+
+export {
+  Checkbox,
+  Radio
+};
