+++ src/choices/src/components/styled/index.js
@@ -1,0 +1,13 @@
+// @ flow
+import CheckboxIndicator from './CheckboxIndicator';
+import ChoiceLabel from './ChoiceLabel';
+import RadioIndicator from './RadioIndicator';
+import { CheckboxInput, RadioInput } from './ChoiceInputs';
+
+export {
+  CheckboxIndicator,
+  CheckboxInput,
+  ChoiceLabel,
+  RadioIndicator,
+  RadioInput
+};
