+++ src/choices/stories/radio.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import Radio from '..';
+import { Radio } from '..';
 import { Card, CardSegment } from '../../layout';
 
 storiesOf('Radio', module)
