+++ src/choices/types.js
@@ -1,0 +1,8 @@
+// @flow
+type IndicatorProps = {
+  checked :boolean;
+};
+
+export type {
+  IndicatorProps
+};
