+++ src/index.js
@@ -16,13 +16,12 @@ export {
 };
 export { default as Banner } from './banner';
 export { default as Button } from './button';
-export { default as Checkbox } from './checkbox';
 export { default as Label } from './label';
 export { default as Modal } from './modal';
 export { default as Overlay } from './overlay';
 export { default as Portal } from './portal';
-export { default as Radio } from './radio';
 export { default as Spinner } from './spinner';
+export { Checkbox, Radio } from './choices';
 export { DatePicker, TimePicker } from './datetime';
 export { Input, TextArea } from './text';
 export {
